AI

Description

AI. Everything to do with the AI gets done in here.

Documentation

data GameTree

Data tpye for game-tress.

Constructors

GameTree

Fields

game_board :: Board

The board at the level.

game_turn :: Col

Whos turn it is.

next_moves :: [(Position, GameTree)]

The next possible moves for this colour, and the trees that they produce.

possiblePositions

Arguments

:: Board
The board.

-> [(Int, Int)]
List of all positions.

Get all the possible positions on the board.

spaces

Arguments

:: Board
The current board.

-> [Position]
List of empty positions.

Gets all the empty positions on the board.

gen

Arguments

:: Board
The current board.

-> Col
The colour of whose turn it is.

-> [Position]
List of free Positions.

Generate List of Positions.

buildTree

Arguments

:: (Board -> Col -> [Position])
Move generator.

-> Board
board state.

-> Col
player to play next.

-> GameTree
resulting game tree.

Given a function to generate plausible moves (i.e. board positions) for a player (Col) on a particular board, generate a (potentially) infinite game tree.

addMovesToList

Arguments

:: GameTree
The game tree with the boards possible moves.

-> [(Int, Position)]
The current list of moves.

-> [(Int, Position)]
The appended list of moves.

Add moves to the current list of possible moves that have been evaluated.

treeTraverse

Arguments

:: Int
The maximum depth to traverse the tree.

-> Int
The current depth.

-> [(Int, Position)]
The list of possible moves.

-> GameTree
The current Game-tree.

-> [(Int, Position)]
The list of evaluated moves.

Going through teh tree and returning a list of evaluated moves.

getBestMove

Arguments

:: Int
Maximum search depth.

-> GameTree
Initial game tree.

-> Position
The best moves available.

Get the best next move from a (possibly infinite) game tree.

maximumFromList

Arguments

:: [(Int, Position)]
The list of evaluated moves.

-> (Int, Position)
The best possible move.

Sorts the list of evaluated moves, and returns the best one.

updateWorld

Arguments

:: Float
Time since last update.

-> World
Current world state.

-> IO World
Returns the IO world, either unchanged, or with the move that the AI has made.

Checks if the game has been won, if not then if it is the AIs turn then it makes a move.