Board

Description

Board. All actions to do with the board the player can do.

Documentation

data Col

Data type representing the turn.

Constructors

Black

White

Empty

other :: Col -> Col

Function to swap the colour to the opposite.

type Position = (Int, Int)

Type specifying what a Position is.

data Board

A Board is a record containing the board size (a board is a square grid, n * n), the number of pieces in a row required to win, and a list of pairs of position and the colour at that position.

Constructors

Board

Fields

size :: Int

The size of the board.

target :: Int

The target of pieces to get in a line.

pieces :: [(Position, Col)]

All the pieces on the board.

initBoard :: Board

The initial Board used to reset the game to defaults.

data World

Overall state is the board and whose turn it is, plus any further information about the world.

Constructors

World
If a hint has been asked for. deriving(Show)

Fields

board :: Board

The board of the game.

turn :: Col

Whos turn it currently is.

won :: Bool

If ther game has been won.

winner :: Col

Who the winner is.

hint :: Bool

pause :: Bool

timer :: Float

time :: Bool

pics :: Bool

squares :: Picture

counterOne :: Picture

counterTwo :: Picture

makeMove

Arguments

:: Board
The current Board.

-> Col
Who's turn it currently is.

-> Position
The Positon of the move the player wants to make.

-> Maybe Board
Either the updated Board or Nothing.

Play a move on the board; return Nothing if the move is invalid (e.g. outside the range of the board, or there is a piece already there).

restartGame

Arguments

:: World
The current World.

-> World
The newly reset world.

Restart and reset the game to its default settings.

changeImage

Arguments

:: World
The current world

-> World
The new updted world

Swap the picture being displayed on the screen between bitmap images and simple gloss images.

increaseBoardSize

Arguments

:: World
The current World.

-> World
The newly increased size world.

Incearse the number of squares on the game board.

decreaseBoardSize

Arguments

:: World
The current World.

-> World
The newly decreased size world.

Decrease the number of squares on the game board.

increaseLineSize

Arguments

:: World
The current World.

-> World
The newly increased target world.

Increase the target line length to get to win the game, onlt if the curretn line length is '3'.

decreaseLineSize

Arguments

:: World
The current World.

-> World
The newly decreased target world.

Decrease the target line length to get to win the game, onlt if the curretn line length is '5'.

undoMove

Arguments

:: World
The current World.

-> World
The world without the most two most recent moves.

Undo the players last move, this also undos the AI's most recent move.

checkPositionOnBoard

Arguments

:: Board
The game Board.

-> Position
The position being checked.

-> Bool
True or False for if it is a position on the board.

Checks if a position is a valid position on the board.

checkPositionForPiece

Arguments

:: Position
The Position being checked.

-> Board
The current board.

-> Bool
True or False if there is a piece or not.

Check if a Position passed in has a pice on it or not.

positions

Arguments

:: Board
Current Board.

-> [(Int, Int)]
List of Popsitions on board.

Generates a list of all the Positions for the baord.

getPositionColor

Arguments

:: Board
The current board.

-> Position
Tye position being checked.

-> Col
The colour of the piece, or Empty

If there is a piece on the position, it returns the colour of the piece.

checkWon

Arguments

:: Board
The current board.

-> Maybe Col
Returns Nothing if neither player has won yet. Returns 'Just c' if the player c has won.

Check whether the board is in a winning state for either player.

checkEast

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the East direction for the target length of lines.

checkWest

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the West direction for the target length of lines.

checkMiddleRow

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the Middle Horizontal position for the target length of lines.

checkNorth

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the North direction for the target length of lines.

checkSouth

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the South direction for the target length of lines.

checkMiddleColumn

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the Middle Vertical position for the target length of lines.

checkNorthEast

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the North East direction for the target length of lines.

checkSouthWest

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the South West direction for the target length of lines.

checkNorthWest

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the North West direction for the target length of lines.

checkSouthEast

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the South East direction for the target length of lines.

checkMiddleDiagonalOne

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the Middle Diagonal Position for the target length of lines.

checkMiddleDiagonalTwo

Arguments

:: Board
The current board.

-> Position
The last move made onto the board.

-> Col
The colour of the last move made on the board.

-> Int
The target length.

-> Bool
True if a line has been found, otherwise false.

Checks in the other Middle Diagonal Position for the target length of lines.

checkAbove

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line North of the Piece.

checkBelow

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line South of the Piece.

checkLeft

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line West of the Piece.

checkRight

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line East of the Piece.

checkUpperLeft

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line North West of the Piece.

checkUpperRight

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line North East of the Piece.

checkLowerLeft

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line South West of the Piece.

checkLowerRight

Arguments

:: Position
The last move made on the board.

-> Col
The colour of the last move made.

-> Board
The current board.

-> Bool
True if two are found, otherwise False.

Checking if there are two pieces in a line South East of the Piece.

checkWhatColor

Arguments

:: (Position, Col)
The Position and colour of the piece at that position.

-> Board
The current board.

-> Col
The colour being compared against.

-> Bool
True or False if the colours are the same.

Compares the colour of a piece to that of the colour passed in.

checkForTwoInARow

Arguments

:: [(Position, Col)]
A list of all the pieces and colours of those pieces.

-> Col
Whos turn it is.

-> Board
The current board.

-> Bool
True if any two pieces are found next to each other.

Function for checking if there are two pieces of the same colour next to each other.

checkForThreeInARow

Arguments

:: [(Position, Col)]
A list of all the pieces and colours of those pieces.

-> Col
Whos turn it is.

-> Board
The current board.

-> Bool
True if any two pieces are found next to each other.

Function for checking if there are is the target amount of pieces of the same colour next to each other.

checkForDefence

Arguments

:: (Position, Col)
The last move made on the board, the move played by the opposite colour.

-> Col
The current colour of whose turn it is.

-> Board
The current board.

-> Bool
True if the player is one away from winning, otherwise false.

Function for seeing if the oppisite colour is one away from getting the target length of pieces in a line.

getPiecesForColor

Arguments

:: Board
The current board.

-> Col
The colour to search for.

-> [(Position, Col)]
List of Positions and Colour.

Gets all the Positions on the board where the specified colour has pieces.

evaluate

Arguments

:: Board
The current board.

-> Col
The colour to evaluate for.

-> Int
If there is a winning line then 5, if it stops the other colour from winner it gets 4, if there are two pieces in a row of the same colour then 2, otherwise 1.

Evaluate how good the board is for the specified colour.